make useSwitchEvmChainIfNeeded hook, remove unnecessarily calling switchToEvmChainIfNeeded manually
